requirements.txt
python-telegram-bot>=21.5
SQLAlchemy>=2.0
aiosqlite>=0.19
APScheduler>=3.10
python-dateutil>=2.9
pydantic>=2.7
python-dotenv>=1.0
tzdata>=2025.1
greenlet>=3.0